got first SEP-12 test working
